move dedup function to utils
